chore(sinon): in non-embroider builds, sinon needs hax
I'm not actually sure what embroider has to do with this, but the main
thing is that sinon imports from util, which isn't a package that it
declares as a peerDependency, and is a package that comes from
node:util,  so this makes sinon a primarily node package, so we should
remove it from browser use.

Additionally, I think this is an issue:
https://github.com/ef4/ember-auto-import/issues/536
v2 addons can't use `@cached` at the moment, because the polyfill is
incorrect for the v2 addon world.