feat(plugins): simplify working with columns among plugins
There is now a new "columns" query-API that "does the right thing",
based on a "requester" (plugin class).
There is a hierarchy of column behaviors that can be statically known,
and the resolution of columns allows plugins to be independent of other
plugins, while also layering in the correct order, based on if the
plugin defines a columns property.