feat: Implement more intuitive column sorting
In order to make the previously added test pass we need to give
the column reordering plugin access to _all_ columns, not just the
visible ones.

When moving left or right we need to iteratively swap the position
of adjacent columns, continuing until we have swapped with a visible
column.

There are some failing tests for internals of the component on this
commit but they can be fixed up later. I want to see if the new UX
is desirable.