fix: reset all column visibility preferences in a single call
When resetting column visibility preferences, each column is currently reset
separately with it's own call to the preferences service.
This update simplifies the reset by adding the `forAllColumns.delete(key: string)` method.
This clears preferences for the given key for the ColumnVisibility plugin for all columns
in a single call.