fix: column resize preferences save
the queueUpdate was uses requestFrameAnimation to schedule the calls
to the resize columns function.
To ensure that the call to save preferences is called after the resize
the queueUpdate now takes an optional callback function that is called
after the resize.
This ensures that these pseudo-async events are always performed in a
predictable and correct sequence.