fix: revert sizing plugin forColumns
None of this logic was previously working, because the feature 'columnResizing'
that was added in commit: 7a098784914b087e423f052ec6b15f16345a56ca and being
used to lookup the plugin was incorrect.
Since we were always getting back the table.columns.values, we can revert this
unused logic.